mellytrade_v3/mellytrade-api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.8,<3.0
SQLAlchemy>=2.0,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
pytest>=8.0,<10.0
